issues fixed in brag section and image section

+++ app/Http/Controllers/BragSectionController.php
@@ -53,7 +53,7 @@ class BragSectionController extends Controller
             // Storage::disk('public-storage')->delete($section->section_image);
             // $data['section_image'] = $request->file('section_image')->store('uploads/brag_section', 'public-storage');
             // Delete the old image from storage
-            if($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg") {
+            if($section->section_image && ($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg")) {
                 Storage::disk('public-storage')->delete($section->section_image);
             }
             
@@ -83,7 +83,7 @@ class BragSectionController extends Controller
         $section = BragSection::findOrFail($sectionId);
         
         // Delete the image associated with the section
-        if($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg") {
+        if($section->section_image && ($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg")) {
             Storage::disk('public-storage')->delete($section->section_image);
         }
